build: add Conan as the dependency and build management system
- Add `conanfile.py`, `conan.lock`, `CMakePresets.json` and Conan
  profiles for gcc, clang, msvc and apple-clang (with asan/ubsan,
  tsan, rtsan variants)
- Switch `CMakeLists.txt` from CPM to `find_package` for
  dependencies managed by Conan (nlohmann_json, SndFile, zstd,
  xxHash, fmt, spdlog, Boost, au, GTest, benchmark, TBB, etc.)
- Add `tools/conanrun_wrapper.sh` to source Conan runtime env under
  gdb while re-applying sanitizer options
- Add `tools/ubsan_suppressions.supp` and expand
  `tools/tsan_suppressions.supp` with Qt/dbus/Wayland suppressions;
  remove obsolete `tools/sanitizer_blacklist.txt`
- Add Conan build instructions to `README.md` and Conan to
  `requirements.txt`
- Fix Qt include paths to use `qqmlintegration.h` for Conan-built Qt
- Drive-by: drop `ZRYTHM_GCC_X86_64_V3_OPTIMIZATIONS` and
  `ZRYTHM_BUILD_WITH_PIPE` options (now handled by profiles), add
  `ZRYTHM_TIME_TRACE` option, and add `-DQT_NO_INT128` workaround
  for GCC 16 with `debug_assert`

+++ src/utils/math_utils.cpp
@@ -21,7 +21,7 @@ calculate_rms_amp (const float * buf, const uint32_t nframes)
       sample = buf[i];
       sum += (sample * sample);
     }
-  return std::sqrtf (sum / ((float) nframes / (float) RMS_FRAMES));
+  return std::sqrt (sum / ((float) nframes / (float) RMS_FRAMES));
 }
 
 /**
